+++ tests/mt/test_query_set.py
@@ -1,0 +1,14 @@
+import pandas as pd
+
+import decoupler as dc
+
+
+def test_query_set(
+    net,
+):
+    ft = set(net[net["source"] == "T1"]["target"])
+    df = dc.mt.query_set(features=ft, net=net, tmin=0)
+    assert isinstance(df, pd.DataFrame)
+    cols = {"source", "stat", "pval", "padj"}
+    assert cols.issubset(df.columns)
+    df = dc.mt.query_set(features=ft, net=net, n_bg=None, tmin=0)